Fixed a bug in MedianFilter and rewrote its reference files for image_test.py.
The median filter should always use an odd kernel size.  Its argument is
an integer radius, but the 2*r+1 needs to be passed to the vtk routine.
It was using 2*r before.  vtk documentation is silent on the issue.

Uncommented the blur and normalize tests, which seem to have been
suffering from a typo in image_test.py, and not actual problems
with the tests.

Added better output when OOFImage::compare fails.