Updated tests for pinned nodes, added some missing features, and moved some code that was in confusing locations.
The menu commands for pinning nodes are now all in OOF.Skeleton.PinNodes,
instead of being split between there and OOF.Graphics_X.Toolbox.Pin_Nodes.
The menu command arguments just contain node positions, not mouse
click details.  This is consistent with the way node selections are
now treated.

Nodes of selected faces can now be pinned.

PinnedNodeSet and PinnedNodeSelection had been stranded in
skeletonnode.py after the vaguely related node selection code had
been moved elsewhere.  I deleted that file and put the code into
skeletonselectable.py because it builds on base classes there.  A
case could be made for putting it in a separate file.

Some menu callbacks that had been in pinnodesmodifier.py are now in
pinnodesmenu.py.  This is consistent with the layout of the skeleton
selection code.

Changed some static functions in cskeletonnode2.C so that they
take advantage of C++11 and return STL containers instead of
going out of their way to avoid doing that.  Added to the comment
about these functions being in the wrong place and being redundant
with code elsewhere.  I'll fix that next.