Various small changes.
Fixed a bug in GhostGfxWindow.findClickedCellCenterMulti that
was crashing if there were no clickable layers in the graphics
window.

Removed the underscores from the names of the toolboxes
whose names had underscores in them.

Added the super-cautious destructor to SubDimensionalSet in
edgeset.py, because one of the tests was raising errors on
exit.  This fixed the symptom but makes me wonder if there's
a memory leak.  Added a TODO.