Fixed a couple bugs in GhostGfxWindow.findClickedCellCenterMulti and PixelElementSelectorGUI.
Selecting elements based on the category of the clicked pixel is
weird, but I've decided not to fix it.  It now works as it is
supposed to, but the question is what it should do.   Use the
dominant pixel of the clicked element, or the clicked pixel?
It uses the pixel, and now doesn't crash if there is no pixel.

findClickedCellCenterMulti wasn't releasing the graphics lock
if there were no suitable layers to click, which was hanging
the program if the above selection method failed.