Fixed a bug in the skeleton info toolbox GUI.
The toolbox was evaluating each mouse click in all modes (element,
face, segment, and node) so that switching modes didn't lose
information.  That raised errors when a click wasn't interpretable
in a mode, even when mode wasn't current, so clicks weren't processed
unless they happened to be valid in all modes.  Now the findClickedXXXX
methods in GhostGfxWindow have an optional "warn" argument that can
be set to False to suppress the errors.