Improvements to abaqus output.
Abaqus output from skeletons and meshes now lists nodes in
a predictable order, so that the tests are repeatable.

Added FEMesh::getNodeIndexMap, which is used in mesh abaqus output
to create a std::map that translates node positions to abaqus node
indices.  Also added a typemap to convert it to a python dict
that maps Points to indices.  This speeds up one of the slow parts
of abaqus file creation by a factor of 7.

Added a TODO in common/typemaps.swg about a typemap that doesn't
seem to do what it says it does.  Maybe it's not used.

Added a reset() method to the timer, for debugging and optimizing.