Sped up mesh abaqus output vastly by moving the node index lookup to C++.
Instead of creating a std::map<Coord, int> in C++ and passing
it to Python, now the map is used in C++. It's wrapped in a swigged
class that accesses it from Python, but the lookup is performed
in C++.  Even with the swig overhead on every call, this is much
faster than doing the lookup in Python.  I don't understand why...

Added comments to the comment in typemaps.swg, indicating that
the Coord* and Coord& typemaps that converted returned Coords to
Points are commented out.