Image layers now become transparent when the box voxel selection widget is active.
This prevents the widget from becoming invisible inside the image.
The image layers catch an "autodim" switchboard signal, which is
sent by the RectangularPrismSelectorGUI.  Other bulk opaque layers,
such as skeleton material displays or contour plots, should catch
the same signal.

The degree to which the layers become transparent should be settable
by the user.