When the voxel selection box widget is visible, SolidFillMeshDisplays are automatically made translucent.
This extends to SolidFillMeshDisplays what was already done for
BitmapDisplayMethods.

Changed the switchboard signal that controls the process from "autodim"
to "autoopacity", and instead of passing a bool, it passes a float
that is the factor by which the native opacity of the layer should
be changed.  This lets the factor be a parameter of the layer that's
requesting the change, which makes more sense.  The factor can now
be set in the VoxelRegionSelectionSelectionDisplay's parameters.

Commented out some destroy() methods in engine.IO.displaymethods
that were doing nothing but calling the superclass's destroy()
method, which doesn't have to be done explicitly.  Maybe they used
to do something useful.