Fixed some bugs related to the clipping plane widget.  The widget for new planes has a more helpful default position.
The Viewer toolbox sets the default parameters for new clipping
planes so that the plane is 1/4 of the way across the microstructure
and perpendicular to the *screen's* x axis.  This makes it less likely
that the plane clips the whole microstructure, and also makes it likely
that both the plane and arrow parts of the widget are visible.

Fixed an error in PlaneAndArrowLayer::set_normal, which wasn't
checking for the cases in which the cross product yielded a zero
vector.  This was causing the initial plane (before user interaction)
to be drawn in the wrong place unless the normal was +x.

Fixed an error in CUnitVectorDirection::CUnitVectorDirection(const Coord*)
which wasn't initializing its data properly.  Apparently this hadn't
been used anywhere before.  Added a constructor that takes a const Coord&.

Simplified a bunch of code in the PlaneAndArrowLayer by taking
advantage of the conversion of Coords to and from double*s, and the
vtk function variants that return double*s and accept double*
arguments.

Changed the return type of some PlaneAndArrowMethods that
were returning Coord*s, so that they now just return Coords.
This doesn't save much computationally, but simplifies the
code and is less error prone.  It adds a hidden copy in the
swig-generated code.

Added versions of dot, cross, and norm2 for Coord3Ds that take
pointers, so that they can be easily called from Python using the
Coord *Point typemap.