Working on modifying setup.py and related files so that oof3d can be built by MacPorts.
Added an option --skip-install-name-tool option to the installation
step because (I think) macports does this itself.

setup.py takes a --portdir argument that tells it to look in something
other than /opt/local for dependencies installed with macports.
This is necessary because when we use macports to build a standalone
distributable package, installng into /opt/local could cause
conflicts.