Added --dest-dir as an argument to setup.py.
I had been misunderstanding how DESTDIR should be used.
install_name_tool is run as if the files are being installed
under the --prefix argument, but they are actually installed
into --dest-dir, which is the equivalent of DESTDIR in a Makefile.
It's set by MacPorts and ignored by everybody else.

make_dist now prints some checksums, which have to be copied
into the Portfile.