Trying to use --install_name when linking shared libraries.
install_name is a file name in the final installation directory, not
the directory where the files are actually written, because MacPorts
writes them in a staging area.  Normally distutils build commands
don't know the installation prefix, so I added a --prefix argument
to build and build_shlib.  That's ugly because it requires it to
be set during the build phase even when not doing a staged installation.