Cleaning up the macports build, which may be mostly working.
Removed --skip-install-name-tool, since install_name is now set in
the distutils link step by build_shlib.  Deleted the code in
install_shlib that was running install_name_tool.

Added --prefix as an option for the distutils build stage.  It's
needed so that --install_name can be passed to the linker.
If build and install are performed together, it's ok if --prefix
is provided only as an install option.  build will copy it from
install.