Commented out some incorrect and unused code in tutorialsGUI.py.
It was trying to get a directory containing images used in the
tutorials, but it was using os.getcwd(), which returns the wrong
directory and also fails if the program is started without a
current directory (eg, cd dir; rm -r ../dir; oof3d).  I just
commented out all of the image code.