Some simplifications and cleanup.
Removed the --skip-install-name-tool arg for setup.py. It's not
needed because we now use --install-name correctly.  The oof_install_lib
class isn't needed anymore, for the same reason.

Attempted to add the test files to the MacPorts distribution, but
it's not working yet.

Removed the --dest-dir argument to install_shlib.  The distutils
--root argument already does what --dest-dir was doing.