The regression tests can now be run by calling 'oof3dtest' in any directory.
oof3dtest is installed into bin alongside oof3d by setup.py.

Added a function OOFglobals() to utils.py.  It returns the contents
of the oof namespace.  This allows regression.py to work whether
or not its __main__ is the same as the oof namespace.

Changed the text of the syntax error raised by the syntaxerror test
in fundamental_test.py.  Because the test files are now installed
by setup.py, syntaxerror.py is precompiled at installation time,
and the syntax error is visible to the users.  So now it explains
that it's intentional.

Deleted the obsolete cvs tags from the toplevel oof3d script.