Fixes to allow building on RHEL 7.5.
File name expansion in the package_data arg in setupargs is reported
not to work on RHEL 7.5 when the names contain two wildcards, so
now all the data files are found by os.walk() and listed explicitly
in package_data.  Any subdirectory named *_data in TEST3D is included
completely.

RHEL 7.5 provides <strstream> but not <strstream.h> or <sstream>,
so the HAVE_SSTREAM test in setup.py and the #ifdef SSTREAM code
in various places in SRC wasn't working.   Added HAVE_STRSTREAM.
Removed an import of <strstream.h>/<sstream> where it wasn't needed.

WTF, RH?