Added another category for message window output, "data".
"Data" differs from "report" in that data is the end result of
a command, and you might want to see it all by itself so that it
can be cut and pasted somewhere else.  "Report" includes all sorts
of status reports that are incidental to the actual result.

Also, the order of the buttons for the various output categories
in the Message Window is now more sensible.
They're in the same order as common.IO.reporter.messageclasses,
instead of the opposite order.