Enhanced checkCategoryVolumes, and menu item to call it.
CSkeleton.checkCategoryVolumes computes the volume of each element
by summing the volumes of each category within the element and
compares it to the geometrical volume.  It also computes the total
volume of each category in the Microstructure by summing over all
elements, and compares it to the known number of voxels in each
category.

OOF.Help.Debug.Check_CategoryVolumes prints the relative error in
the volume of each element averaged over all elements, the rms
variation in the error, and the maximum error.  It also prints the
relative error in each category.