Moved CRectangularPrism and ICRectangularPrism from geometry.* to cprism.*.
These changes make the voxel set boundary code more modular, so
that it can be distributed independently.

cprism.* contains template code that is explicitly used by voxelsetboundary.*.
The template parameter is a Coord3D-like class.
geometry.* is OOF-specific.

Fixed a test that was broken when checkCategoryVolumes was changed
to return error data instead of comparing the error to a tolerance.