Voxel set intersection code is now independent of the rest of OOF3D.
Moved protonode subclasses from VSB/vsb.h to VSB/protonode.h
just because the file was getting too big.  It doesn't simplify
compilation at all.

There's a new templated 3D array class in VSB/array3d.h so that the
VSB calling code doesn't have to provide an array class.  VSB/array3d.h
is simpler than common/array.h, and possibly could be used instead
of it.

Instead of having two oriented plane classes or making non-OOF VSB
users provide their own, the VSB oriented plane class is now used
by the rest of OOF3D, via a typedef in commmon/geometry.h.

common/voxelsetboundary.* defines a VoxelSetBoundary class, which is
an OOF3D-specific wrapper for the templated VoxelSetBdy class in
VSB/vsb.h.   Currently a VoxelSetBoundary contains a pointer to
a VoxelSetBdy, but it might be better to use inheritance instead.
That's the next thing to look at.

Some of the work that used to be done in CSkeleton::buildVSBs()
is now done in a function, buildVSB(), that is called by the
VoxelSetBoundary constructor.  buildVSB is defined in VSB/vsb.h.

The debugging and plotting methods in vsb.h can't do the OOF3D-specific
things that they used to do, so they're now defined as template methods
and take an argument that serves as a callback function.

There is still some cleaning up to do.  Subregions are being passed
as arguments where they don't have to be, for example.