Monkeypatch to disutils.command.build_py to fix a Red Hat problem.
Red Hat/CentOS still uses Python 2.7.5.  setup.py in 2.7.6 and earlier didn't
support directory trees in package_data, which we require in order to
install the reference files for the test suite.  This change copies
build_py.find_data_files() from 2.7.x (x >= 7) and monkeypatches it in
if the python version is less than 2.7.7.