Rewrote GhostOOFCanvas::findClickedActors so that it works with VTK 8.2.0.
findClickedActors now uses one call to vtkPicker instead of many calls
to vtkPropPicker.   I don't know why vtkPropPicker wasn't working.  I
tried to reproduce the problem in a simple code, and couldn't.

Debugging lines are still in.  findClickedPositionOnActor is still
broken.  I'm commiting this now so that the debugging lines can
be restored later if necessary.