GhostOOFCanvas::findClickedPositionOnActor now uses vtkPicker instead of vtkPropPicker.
This slightly breaks ctrl-clicking on the plane-and-arrow layer
when editing a clipping plane.  The arrow is always being selected
instead of the plane, and the initial motion of the arrow is jumpy.
However, it doesn't fail completely the way it did when using
vtkPropPicker.

Removed debugging lines, and otherwise cleaned up some code.
Commented out findClickedActor, which wasn't being used.
findClickedActors and findClickedPositionOnActor now share some
code.