Fixed problems with clipping plane editing when using vtk 8.2.0.
The vtkPropPicker in vtk 8.2.0 doesn't seem to work properly, although
I was unable to get a test program to fail.  Replacing it with vtkPicker
didn't work, because vtkPicker just uses bounding boxes, and wasn't
correctly determining when the arrow in the PlaneAndArrow layer was
selected.  Using vtkCellPicker instead might be overkill, but it does
seem to work.

The Viewer toolbox gui is now more careful about resetting self.operation
if the click fails. Also added a minor TODO about making self.operation
a class.

Deleted GhostGfxWindow.findClickedActor and related methods in
GhostOOFCanvas.  It wasn't being used, and was nearly redundant
with findClickedActors in any case.