Fixed some more slowness in mesh abaqus output.
Nodes for face boundary definitions were being stored in a C++
set which was converted to a python OrderedSet in a swig typemap,
and then loaded into a python list, which was then sorted.  This
was dumb.  Now the C++ set is converted directly into a python
list by the typemap.

Also eliminated the extra loops over elements that were being done
when creating element sets for each material.

Also added progress bars to the parts of the abaqus output routine
that might be slow.