Converted the jpeg files used in image and pixel group tests to png.
The system jpeg library on the Mac is now different from that on
Linux, and tests that rely on the exact RGB values of voxels will
fail on one platform or the other (unless vtk on the Mac is compiled
so that it uses its own jpeg library, instead of the system library).
This is an attempt to avoid the problem entirely by using png files
instead of jpeg.