Starting to add a simple, fast, and wrong alternative homogeneity calculation to compare to VSB.
This version doesn't compile on the Mac, although none of the changes
have anything to do with the compilation errors.  I'm committing it
to move it to Linux to see what happens there.

Addded a HomogeneityAlgorithm enum that controls which method is used.

Added CachedValue::clear() so that cached homogeneity values can
be forced to be recomputed when the homogeneity algorithm is changed.

Added a menu command to change the homogeneity algorithm.

Added resetHomogeneity methods to SkeletonContext, CSkeleton,
and CSkeletonElement.

CSkeletonElement::c_homogeneity now calls either c_homogeneity_robust
or c_homogeneity_fast, depending on the enum setting.