Use SHLIB_EXT from setup.py config vars to set the name of the oof3d .so files.
SHLIB_EXT is .so in most places, but when using python-dbg it's _d.so.
When building oof3d with python-dbg, the .so files it creates automatically
are created with _d.so, but it needs to be told explicitly to link to *_d.so
instead of *.so.
addOOFlibs() in setup.py can be used in DIR.py file to generate the
correct lib name for linking.